feat: Add react-router-dom package and update dependencies
This commit adds the react-router-dom package to the project, allowing for client-side routing in the React app. It also updates the package.json and package-lock.json files to include the new dependency. This change was made to enable navigation between different pages in the application. Inserted SmoochSans-Medium , SmoochSans-Light , Raleway-Medium Fonts